#161 Preparing handling of TypeInternal via MR JAR;
* Moving MirrorTypeInternal and ReflectType into their own source files
* Creating TypeInternalProviders class
 - returns all three providers under JDK 8
 - returns just reflection and mirror provider under JDK 9
* Removing reflective lookup of TreeTypeInternal
